--- components/Database.js
@@ -1,158 +1,0 @@
-// Videos Data
-
-export const videoData = [
-  {
-    video: require('../../storage/videos/video1.mp4'),
-    postProfile: require('../../storage/images/post1.jpg'),
-    title: 'Ram_Charan',
-    description: 'Feel the buity of nature',
-    likes: '245k',
-    isLike: false,
-  },
-  {
-    video: require('../../storage/videos/video2.mp4'),
-    postProfile: require('../../storage/images/post2.jpg'),
-    title: 'The_Groot',
-    description: "It's a tea time",
-    likes: '656k',
-    isLike: false,
-  },
-  {
-    video: require('../../storage/videos/video3.mp4'),
-    postProfile: require('../../storage/images/post3.jpg'),
-    title: 'loverland',
-    description: 'Feel the buity of nature',
-    likes: '243k',
-    isLike: false,
-  },
-  {
-    video: require('../../storage/videos/video4.mp4'),
-    postProfile: require('../../storage/images/post4.jpg'),
-    title: 'mr. bean',
-    description: 'How cute it is !!',
-    likes: '876k',
-    isLike: false,
-  },
-];
-
-// Friends Profile Data
-
-export const FriendsProfileData = [
-  {
-    name: 'Ram_Charan',
-    accountName: 'Ram Charan',
-    profileImage: require('../../storage/images/profile4.jpg'),
-    posts: 56,
-    followers: 6542,
-    following: 43,
-    follow: false,
-  },
-  {
-    name: 'The_Tom',
-    accountName: 'Tom',
-    profileImage: require('../../storage/images/profile5.jpg'),
-    posts: 24,
-    followers: 1253,
-    following: 64,
-    follow: false,
-  },
-  {
-    name: 'live_long',
-    accountName: 'Live Long',
-    profileImage: require('../../storage/images/profile2.jpg'),
-    posts: 21,
-    followers: 7886,
-    following: 32,
-    follow: false,
-  },
-  {
-    name: 'the_real_hero',
-    accountName: 'Ram Charan',
-    profileImage: require('../../storage/images/post1.jpg'),
-    posts: 123,
-    followers: 64253,
-    following: 32,
-    follow: false,
-  },
-  {
-    name: 'the_jerry',
-    accountName: 'The Jerry',
-    profileImage: require('../../storage/images/post2.jpg'),
-    posts: 56,
-    followers: 6542,
-    following: 43,
-    follow: false,
-  },
-  {
-    name: 'angry_bird',
-    accountName: 'Angry Bird',
-    profileImage: require('../../storage/images/post3.jpg'),
-    posts: 452,
-    followers: '564k',
-    following: 31,
-    follow: false,
-  },
-  {
-    name: 'mr_bean',
-    accountName: 'Mr Bean',
-    profileImage: require('../../storage/images/post4.jpg'),
-    posts: 543,
-    followers: '435k',
-    following: 22,
-    follow: false,
-  },
-  {
-    name: 'the_Jd',
-    accountName: 'Mr JD',
-    profileImage: require('../../storage/images/post5.jpg'),
-    posts: 234,
-    followers: '763k',
-    following: 332,
-    follow: false,
-  },
-  {
-    name: '_don_',
-    accountName: 'Don',
-    profileImage: require('../../storage/images/post6.jpg'),
-    posts: 111,
-    followers: 11223,
-    following: 1,
-    follow: false,
-  },
-  {
-    name: 'black_white',
-    accountName: 'blackWhite',
-    profileImage: require('../../storage/images/post7.jpg'),
-    posts: 121,
-    followers: 43213,
-    following: 21,
-    follow: false,
-  },
-  {
-    name: 'iron_man',
-    accountName: 'Iron Man',
-    profileImage: require('../../storage/images/post8.jpg'),
-    posts: 432,
-    followers: '987k',
-    following: 24,
-    follow: false,
-  },
-  {
-    name: 'funny_videos',
-    accountName: 'Funny Video Official',
-    profileImage: require('../../storage/images/post9.jpg'),
-    posts: '1.2k',
-    followers: '1.2M',
-    following: 12,
-    follow: false,
-  },
-  {
-    name: 'mr_jhon',
-    accountName: 'Mr Jhony',
-    profileImage: require('../../storage/images/post10.jpg'),
-    posts: 533,
-    followers: 64322,
-    following: 123,
-    follow: false,
-  },
-];
